sourcemodule: Move id helpers to subfolder io-id
