+++ zera-modules/sourcemodule/CMakeLists.txt
@@ -31,6 +31,7 @@ find_Package(VfCpp REQUIRED)
 file(GLOB SOURCES 
     ${PROJECT_SOURCE_DIR}/src/*.cpp
     ${PROJECT_SOURCE_DIR}/src/device/*.cpp
+    ${PROJECT_SOURCE_DIR}/src/io-ids/*.cpp
     ${PROJECT_SOURCE_DIR}/src/json/*.cpp
     ${PROJECT_SOURCE_DIR}/src/module/*.cpp
     )
@@ -38,6 +39,7 @@ file(GLOB SOURCES
 file(GLOB HEADER
     ${PROJECT_SOURCE_DIR}/src/*.h
     ${PROJECT_SOURCE_DIR}/src/device/*.h
+    ${PROJECT_SOURCE_DIR}/src/io-ids/*.h
     ${PROJECT_SOURCE_DIR}/src/json/*.h
     ${PROJECT_SOURCE_DIR}/src/module/*.h
     )
